Move noise machine into its own window with mixing-board UI
